+++ Sources/PaymentButtons/PaymentButton.swift
@@ -242,11 +242,6 @@ public class PaymentButton: UIButton {
         } else {
             containerView.layer.cornerRadius = edges.cornerRadius(for: containerView)
         }
-
-        if color != .gold {
-            containerView.layer.borderWidth = 1
-            containerView.layer.borderColor = color.borderColor.cgColor
-        }
     }
 
     // MARK: - Utility